+++ lib/nordigen-ruby.rb
@@ -27,8 +27,15 @@ module Nordigen
             @requisition = RequisitionsApi.new(client=self)
         end
 
-        def request
+        def request(params)
             # HTTP client request
+            parameters = {}
+            params.each do |key, value|
+                if value
+                    parameters[key] = value
+                end
+            end
+
             @request ||= Faraday.new do |conn|
                 conn.url_prefix = BASE_URL
                 conn.headers = @@headers