+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [2.0.0] - 2022-05-05
 
 - Use JSON response instead of OpenStruct
-- Add country filter
+- Add `date_to`, `date_from` arguments to filter transactions
 
 
 ## [1.0.1] - 2022-01-13